frontend/requirements.txt
streamlit==1.43.2
requests==2.32.3
python-dotenv==1.0.1
dataclasses==0.6
plotly==6.0.1